# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HF3 Backend",
        "description": "A backend project made in HF3",
        "contact": {
            "email": "kristian@binau.me"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product-types": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of product types",
                "description": "Returns list of product types",
                "operationId": "ProductTypeController.index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/product-types/{productTypeId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product type",
                "description": "Returns Get product type",
                "operationId": "ProductTypeController.show",
                "parameters": [
                    {
                        "name": "productTypeId",
                        "in": "path",
                        "description": "Product Type Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/regions": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get list of regions",
                "description": "Returns list of regions",
                "operationId": "RegionController.index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/regions/{regionId}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get region",
                "description": "Returns Get region",
                "operationId": "RegionController.show",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "description": "Region Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/region/{regionId}/sub-regions": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get list of sub regions in region",
                "description": "Returns list of sub regions in region",
                "operationId": "RegionSubRegionController.index",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "description": "Region Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/sub-regions": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get list of sub region",
                "description": "Returns list of sub region",
                "operationId": "SubRegionController.index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/sub-regions/{subRegionId}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get sub region",
                "description": "Returns Get sub region",
                "operationId": "SubRegionController.show",
                "parameters": [
                    {
                        "name": "subRegionId",
                        "in": "path",
                        "description": "Sub Region Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
    "security": [
        []
    ]
}